requirements.txt
# Core dependencies for BIOIMAGIN Wolffia Analysis System

# Image processing
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Machine learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Web framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0

# Additional utilities
python-dateutil>=2.8.0
# Update the requirements.txt by adding:
cellpose>=2.2.3
torch>=1.12.0
opencv-python-headless>=4.8.0  # Use headless version for server